Move inteltool log parser code to separate module
